Define and conclude incomig and outgoing helpers, conclude TDD for now about incoming and ourgoing and update migrations change float to decimal

+++ database/migrations/2022_06_09_154914_create_taxes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('taxes', function (Blueprint $table) {
             $table->id();
             $table->string('name')->unique();
-            $table->float('percentage');
+            $table->decimal('percentage',2,2);
             $table->timestamps();
             $table->engine = 'InnoDB';
         });
