+++ database/migrations/2022_06_09_155338_create_outgoings_table.php
@@ -20,9 +20,9 @@ return new class extends Migration
             $table->unsignedBigInteger('actor_id');
             $table->foreign('actor_id')->references('id')->on('actors')->onDelete('cascade');
             $table->json('products');
-            $table->float('subtotal');
-            $table->float('taxes');
-            $table->float('total');
+            $table->decimal('subtotal', 10, 2);
+            $table->decimal('taxes', 10, 2);
+            $table->decimal('total', 10, 2);
             $table->timestamps();
             $table->engine = 'InnoDB';
         });
